Seperating worklog services by staff and date ranges

+++ src/main/java/com/fileee/payrole/repository/WorklogRepository.java
@@ -11,11 +11,12 @@ import com.fileee.payrole.beans.Staff;
 import com.fileee.payrole.beans.Worklog;
 
 @Repository
-public interface WorklogRepository extends JpaRepository<Worklog, Long>{
+public interface WorklogRepository extends JpaRepository<Worklog, Integer>{
 
 	@Query(value = "select w from Worklog w WHERE (w.workingDate BETWEEN :fromDate AND :toDate) AND w.staff = :staff ORDER BY w.workingDate")
 	List<Worklog> findWorklogByStaffInDateRange(Date fromDate, Date toDate, Staff staff);
 		
 	@Query(value = "select w from Worklog w WHERE w.staff = :staff ORDER BY w.workingDate")
 	List<Worklog> findWorklogsByStaff(Staff staff);
+
 }
